#6 Fix: feature/5 병합 충돌 오류 해결

+++ src/main/java/umc/study/domain/Review.java
@@ -27,7 +27,7 @@ public class Review extends BaseEntity {
     private String imageUrls;
 
     @Column(nullable = false)
-    private Integer score;
+    private Float score;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
